# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bolsa de Empleo",
        "description": "Sistema de gestión de ofertas de empleo y seguimiento de candidatos.",
        "contact": {
            "email": "admin@centro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
...
        "/api/titulos": {
            "get": {
                "tags": [
                    "Títulos"
                ],
                "summary": "Listar todos los títulos",
                "operationId": "67e5169604282b92da35b9bb185e87ed",
                "responses": {
                    "200": {
                        "description": "Lista recuperada con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "titulo": {
                                                        "type": "string"
                                                    },
                                                    "estado": {
                                                        "type": "string"
                                                    },
                                                    "nivel": {
                                                        "type": "string"
                                                    },
                                                    "familia": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Títulos"
                ],
                "summary": "Crear nuevo título",
                "operationId": "9e95c555b4f1cc8d28775e8f298b12fb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "nivel",
                                    "familia",
                                    "centro"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "nivel": {
                                        "type": "integer"
                                    },
                                    "familia": {
                                        "type": "integer"
                                    },
                                    "centro": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/titulos/niveles/listado": {
            "get": {
                "tags": [
                    "Títulos"
                ],
                "summary": "Obtener niveles educativos",
                "operationId": "bc96f957be681037d0753d42d1c3cd6d",
                "responses": {
                    "200": {
                        "description": "Niveles recuperados."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/titulos/{titulo}": {
            "get": {
                "tags": [
                    "Títulos"
                ],
                "summary": "Ver detalle de un título",
                "operationId": "d5aa77926109a10d8b5534f3a04de3ae",
                "parameters": [
                    {
                        "name": "titulo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle obtenido."
                    },
                    "404": {
                        "description": "No encontrado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Títulos"
                ],
                "summary": "Desactivar título",
                "operationId": "472b6128e43f9fbd31c95bfb27db89ff",
                "parameters": [
                    {
                        "name": "titulo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Desactivado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Títulos"
                ],
                "summary": "Actualizar título existente",
                "operationId": "1b87e47afd3af002fdb04efa72226977",
                "parameters": [
                    {
                        "name": "titulo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actualizado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/titulos/activos": {
            "get": {
                "tags": [
                    "Títulos"
                ],
                "summary": "Listar solo títulos activos",
                "operationId": "b4da2f8499d782770df4e119ef7d2483",
                "responses": {
                    "200": {
                        "description": "Éxito."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/titulos/demandante": {
            "get": {
                "tags": [
                    "Títulos-Demandante"
                ],
                "summary": "Ver mis títulos (Demandante)",
                "operationId": "4f3132256003988643ad399d5f0e4888",
                "responses": {
                    "200": {
                        "description": "Lista."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Títulos-Demandante"
                ],
                "summary": "Vincular títulos a un perfil de demandante",
                "operationId": "a2f4311be37b43c54d41eca89735ad85",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulos": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "centro": {
                                                    "type": "string"
                                                },
                                                "anio": {
                                                    "type": "integer"
                                                },
                                                "cursando": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vinculados."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/titulos/demandante/{id}": {
            "delete": {
                "tags": [
                    "Títulos-Demandante"
                ],
                "summary": "Eliminar un título del demandante",
                "operationId": "49944c87e96dd57f7f2b7d12db25072f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Eliminado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "Authorization",
                "in": "header",
                "example": "Bearer 19|tBfCzyfhD1Ib3VA5HPyHfR6CjrIoyT5VPq7s0mMnb19cdfbe"
            }
        }
    },
...
}